// gasy-car-frontend/src/types/addVehicleType.ts
export interface VehicleFormData {
    // 1️⃣ Identité du véhicule
    titre: string;
    marque: string;
    modele: string;
    annee: number;
    numero_immatriculation: string;
    numero_serie: string;
    proprietaire: string;
    // 2️⃣ Typologie
    categorie: string;
    transmission: string;
    type_carburant: string;
    statut: string | null;
    type_vehicule: "UTILITAIRE" | "TOURISME";
    // 3️⃣ Localisation du véhicule
    adresse_localisation: string;
    ville: string;
    zone: string;
    // 3️⃣ Tarification & caution
    prix_jour: number;
    prix_heure?: number | null;
    limite_heures?: number | null;
    prix_mois?: number | null;
    prix_par_semaine?: number | null;
    remise_par_heure?: number | null;
    remise_par_jour?: number | null;
    remise_par_mois?: number | null;
    devise: string;
    remise_longue_duree_pourcent?: number | null;
    montant_caution?: number | null;

    // 4️⃣ Tarification Province (Optionnelle)
    province_prix_jour?: number | null;
    province_prix_heure?: number | null;
    province_prix_mois?: number | null;
    province_prix_par_semaine?: number | null;
    province_remise_par_heure?: number | null;
    province_remise_par_jour?: number | null;
    province_remise_par_mois?: number | null;
    province_remise_longue_duree_pourcent?: number | null;

    // 5️⃣ Caractéristiques
    nombre_places: number;
    nombre_portes: number;
    couleur: string;
    kilometrage_actuel_km: number;
    volume_coffre_litres: number | null;
    // 5️⃣ Équipements & options
    equipements: string[];
    // 7️⃣ Publication
    est_certifie?: boolean;
    est_disponible: boolean;
    description: string;
    conditions_particulieres: string;
    note_moyenne: string | null;
    nombre_locations: number;
    nombre_favoris: number;
}